#7008 - BE - Manage response code mappings from payment processors (CRUD)
- Adding soft deletes functionality

+++ src/main/java/com/mcmcg/ico/bluefin/persistent/jpa/PaymentProcessorResponseCodeRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mcmcg.ico.bluefin.persistent.PaymentProcessorResponseCode;
 
-public interface PaymentProcessorResponseCodeRepository extends JpaRepository<PaymentProcessorResponseCode, Integer> {
+public interface PaymentProcessorResponseCodeRepository extends JpaRepository<PaymentProcessorResponseCode, Long> {
 
     public PaymentProcessorResponseCode findByPaymentProcessorResponseCode(String paymentProcessorResponseCode);
 }
